Sync epiworld 0.17.0 (complete bubble ties, directed networks); bump to 0.17.0-0 (#194)
* Sync epiworld 0.17.0 (complete bubble ties, directed networks); bump to 0.17.0-0

- Sync inst/include/epiworld with epiworld v0.17.0.
- bubbles() gains `ties = c("existing", "complete")`; "complete" ties every
  pair sharing a bubble while the policy is in force.
- Document directed-tie semantics in agents_smallworld()/agents_from_edgelist().
- Tests for complete ties and directed transmission.
- Bubbles vignette: add a complete-ties scenario.
- NEWS for 0.17.0-0.
- AGENTS.md/CLAUDE.md: agents build, test and document inside the devcontainer.
- Docs regenerated in the devcontainer (roxygen2 8.1.0).


* Add Niko Speybroeck as author (design of the bubbles intervention)


* Address review: vignette wording, regression test for rebuilding after run_multiple()


* Re-sync epiworld 0.17.0 (withdraw Complete bubble ties on run(0)) (#195)


---------

+++ inst/include/epiworld/randgraph.hpp
@@ -123,14 +123,14 @@ inline void rewire_degseq(
         // After swap: p0 will be connected to neighbor_id_11, p1 to neighbor_id_01
         bool would_create_duplicate = false;
         for (auto* n : neighbors_p0) {
-            if (n->get_id() == neighbor_id_11 && n != neighbors_p0[id01]) {
+            if (static_cast<size_t>(n->get_id()) == neighbor_id_11 && n != neighbors_p0[id01]) {
                 would_create_duplicate = true;
                 break;
             }
         }
         if (!would_create_duplicate) {
             for (auto* n : neighbors_p1) {
-                if (n->get_id() == neighbor_id_01 && n != neighbors_p1[id11]) {
+                if (static_cast<size_t>(n->get_id()) == neighbor_id_01 && n != neighbors_p1[id11]) {
                     would_create_duplicate = true;
                     break;
                 }
